requirements.txt
streamlit
scikit-learn
scipy
pandas
numpy
requests
pickle5
thefuzz
python-Levenshtein
python-dotenv